[Refactor] Move strokeShape from LGraphCanvas to draw
